+++ src/stewart_platform/scripts/MBPO/MBPO_main.py
@@ -1,7 +1,6 @@
 import argparse
 import time
 import gym
-import torch
 import numpy as np
 from itertools import count
 
@@ -11,152 +10,12 @@ import os
 import os.path as osp
 import json
 
-from sac.replay_memory import ReplayMemory
-from sac.sac import SAC
-from model import EnsembleDynamicsModel
+from tf_models.replay_memory import ReplayMemory
+# from sac.sac import SAC
 from predict_env import PredictEnv
 from sample_env import EnvSampler
 from tf_models.constructor import construct_model, format_samples_for_training
 
-
-def readParser():
-    parser = argparse.ArgumentParser(description='MBPO')
-    parser.add_argument('--env_name', default="Hopper-v2",
-                        help='Mujoco Gym environment (default: Hopper-v2)')
-    parser.add_argument('--seed', type=int, default=123456, metavar='N',
-                        help='random seed (default: 123456)')
-
-    parser.add_argument('--use_decay', type=bool, default=True, metavar='G',
-                        help='discount factor for reward (default: 0.99)')
-
-    parser.add_argument('--gamma', type=float, default=0.99, metavar='G',
-                        help='discount factor for reward (default: 0.99)')
-    parser.add_argument('--tau', type=float, default=0.005, metavar='G',
-                        help='target smoothing coefficient(τ) (default: 0.005)')
-    parser.add_argument('--alpha', type=float, default=0.2, metavar='G',
-                        help='Temperature parameter α determines the relative importance of the entropy\
-                            term against the reward (default: 0.2)')
-    parser.add_argument('--policy', default="Gaussian",
-                        help='Policy Type: Gaussian | Deterministic (default: Gaussian)')
-    parser.add_argument('--target_update_interval', type=int, default=1, metavar='N',
-                        help='Value target update per no. of updates per step (default: 1)')
-    parser.add_argument('--automatic_entropy_tuning', type=bool, default=False, metavar='G',
-                        help='Automaically adjust α (default: False)')
-    parser.add_argument('--hidden_size', type=int, default=256, metavar='N',
-                        help='hidden size (default: 256)')
-    parser.add_argument('--lr', type=float, default=0.0003, metavar='G',
-                        help='learning rate (default: 0.0003)')
-
-    parser.add_argument('--num_networks', type=int, default=7, metavar='E',
-                        help='ensemble size (default: 7)')
-    parser.add_argument('--num_elites', type=int, default=5, metavar='E',
-                        help='elite size (default: 5)')
-    parser.add_argument('--pred_hidden_size', type=int, default=200, metavar='E',
-                        help='hidden size for predictive model')
-    parser.add_argument('--reward_size', type=int, default=1, metavar='E',
-                        help='environment reward size')
-
-    parser.add_argument('--replay_size', type=int, default=1000000, metavar='N',
-                        help='size of replay buffer (default: 10000000)')
-
-    parser.add_argument('--model_retain_epochs', type=int, default=1, metavar='A',
-                        help='retain epochs')
-    parser.add_argument('--model_train_freq', type=int, default=250, metavar='A',
-                        help='frequency of training')
-    parser.add_argument('--rollout_batch_size', type=int, default=100000, metavar='A',
-                        help='rollout number M')
-    parser.add_argument('--epoch_length', type=int, default=1000, metavar='A',
-                        help='steps per epoch')
-    parser.add_argument('--rollout_min_epoch', type=int, default=20, metavar='A',
-                        help='rollout min epoch')
-    parser.add_argument('--rollout_max_epoch', type=int, default=150, metavar='A',
-                        help='rollout max epoch')
-    parser.add_argument('--rollout_min_length', type=int, default=1, metavar='A',
-                        help='rollout min length')
-    parser.add_argument('--rollout_max_length', type=int, default=15, metavar='A',
-                        help='rollout max length')
-    parser.add_argument('--num_epoch', type=int, default=1000, metavar='A',
-                        help='total number of epochs')
-    parser.add_argument('--min_pool_size', type=int, default=1000, metavar='A',
-                        help='minimum pool size')
-    parser.add_argument('--real_ratio', type=float, default=0.05, metavar='A',
-                        help='ratio of env samples / model samples')
-    parser.add_argument('--train_every_n_steps', type=int, default=1, metavar='A',
-                        help='frequency of training policy')
-    parser.add_argument('--num_train_repeat', type=int, default=20, metavar='A',
-                        help='times to training policy per step')
-    parser.add_argument('--max_train_repeat_per_step', type=int, default=5, metavar='A',
-                        help='max training times per step')
-    parser.add_argument('--policy_train_batch_size', type=int, default=256, metavar='A',
-                        help='batch size for training policy')
-    parser.add_argument('--init_exploration_steps', type=int, default=5000, metavar='A',
-                        help='exploration steps initially')
-    parser.add_argument('--max_path_length', type=int, default=1000, metavar='A',
-                        help='max length of path')
-
-
-    parser.add_argument('--model_type', default='tensorflow', metavar='A',
-                        help='predict model -- pytorch or tensorflow')
-
-    parser.add_argument('--cuda', default=True, action="store_true",
-                        help='run on CUDA (default: True)')
-    return parser.parse_args()
-
-
-def train(args, env_sampler, predict_env, agent, env_pool, model_pool):
-    total_step = 0
-    reward_sum = 0
-    rollout_length = 1
-    exploration_before_start(args, env_sampler, env_pool, agent)
-
-    for epoch_step in range(args.num_epoch):
-        start_step = total_step
-        train_policy_steps = 0
-        for i in count():
-            cur_step = total_step - start_step
-
-            if cur_step >= args.epoch_length and len(env_pool) > args.min_pool_size:
-                break
-
-            if cur_step > 0 and cur_step % args.model_train_freq == 0 and args.real_ratio < 1.0:
-                train_predict_model(args, env_pool, predict_env)
-
-                new_rollout_length = set_rollout_length(args, epoch_step)
-                if rollout_length != new_rollout_length:
-                    rollout_length = new_rollout_length
-                    model_pool = resize_model_pool(args, rollout_length, model_pool)
-
-                rollout_model(args, predict_env, agent, model_pool, env_pool, rollout_length)
-
-            cur_state, action, next_state, reward, done, info = env_sampler.sample(agent)
-            env_pool.push(cur_state, action, reward, next_state, done)
-
-            if len(env_pool) > args.min_pool_size:
-                train_policy_steps += train_policy_repeats(args, total_step, train_policy_steps, cur_step, env_pool, model_pool, agent)
-
-            total_step += 1
-
-            if total_step % args.epoch_length == 0:
-                '''
-                avg_reward_len = min(len(env_sampler.path_rewards), 5)
-                avg_reward = sum(env_sampler.path_rewards[-avg_reward_len:]) / avg_reward_len
-                logging.info("Step Reward: " + str(total_step) + " " + str(env_sampler.path_rewards[-1]) + " " + str(avg_reward))
-                print(total_step, env_sampler.path_rewards[-1], avg_reward)
-                '''
-                env_sampler.current_state = None
-                sum_reward = 0
-                done = False
-                test_step = 0
-
-                while (not done) and (test_step != args.max_path_length):
-                    cur_state, action, next_state, reward, done, info = env_sampler.sample(agent, eval_t=True)
-                    sum_reward += reward
-                    test_step += 1
-                # logger.record_tabular("total_step", total_step)
-                # logger.record_tabular("sum_reward", sum_reward)
-                # logger.dump_tabular()
-                logging.info("Step Reward: " + str(total_step) + " " + str(sum_reward))
-                # print(total_step, sum_reward)
 
 
 def exploration_before_start(args, env_sampler, env_pool, agent):
@@ -238,3 +97,60 @@ def train_policy_repeats(args, total_step, train_step, cur_step, env_pool, model
         agent.update_parameters((batch_state, batch_action, batch_reward, batch_next_state, batch_done), args.policy_train_batch_size, i)
 
     return args.num_train_repeat
+
+
+
+def train(args, env_sampler, predict_env, agent, env_pool, model_pool):
+    total_step = 0
+    reward_sum = 0
+    rollout_length = 1
+    exploration_before_start(args, env_sampler, env_pool, agent)
+
+    for epoch_step in range(args.num_epoch):
+        start_step = total_step
+        train_policy_steps = 0
+        for i in count():
+            cur_step = total_step - start_step
+
+            if cur_step >= args.epoch_length and len(env_pool) > args.min_pool_size:
+                break
+
+            if cur_step > 0 and cur_step % args.model_train_freq == 0 and args.real_ratio < 1.0:
+                train_predict_model(args, env_pool, predict_env)
+
+                new_rollout_length = set_rollout_length(args, epoch_step)
+                if rollout_length != new_rollout_length:
+                    rollout_length = new_rollout_length
+                    model_pool = resize_model_pool(args, rollout_length, model_pool)
+
+                rollout_model(args, predict_env, agent, model_pool, env_pool, rollout_length)
+
+            cur_state, action, next_state, reward, done, info = env_sampler.sample(agent)
+            env_pool.push(cur_state, action, reward, next_state, done)
+
+            if len(env_pool) > args.min_pool_size:
+                train_policy_steps += train_policy_repeats(args, total_step, train_policy_steps, cur_step, env_pool, model_pool, agent)
+
+            total_step += 1
+
+            if total_step % args.epoch_length == 0:
+                '''
+                avg_reward_len = min(len(env_sampler.path_rewards), 5)
+                avg_reward = sum(env_sampler.path_rewards[-avg_reward_len:]) / avg_reward_len
+                logging.info("Step Reward: " + str(total_step) + " " + str(env_sampler.path_rewards[-1]) + " " + str(avg_reward))
+                print(total_step, env_sampler.path_rewards[-1], avg_reward)
+                '''
+                env_sampler.current_state = None
+                sum_reward = 0
+                done = False
+                test_step = 0
+
+                while (not done) and (test_step != args.max_path_length):
+                    cur_state, action, next_state, reward, done, info = env_sampler.sample(agent, eval_t=True)
+                    sum_reward += reward
+                    test_step += 1
+                # logger.record_tabular("total_step", total_step)
+                # logger.record_tabular("sum_reward", sum_reward)
+                # logger.dump_tabular()
+                logging.info("Step Reward: " + str(total_step) + " " + str(sum_reward))
+                # print(total_step, sum_reward)
